Move test classes related to CFG/uCFG
